+++ src/tokens/base.ts
@@ -45,6 +45,7 @@ export enum CoinKey {
   FLOW = 'FLOW', // Flow native
   MON = 'MON', // Monad native
   GUSDT = 'GUSDT', // Stable native (Gas USDT)
+  ZEROG = '0G', // 0G native
 
   // SVM
   SOL = 'SOL',
@@ -67,6 +68,9 @@ export enum CoinKey {
   // Tron
   TRX = 'TRX',
   WTRX = 'WTRX',
+
+  // Stellar
+  XLM = 'XLM',
 
   // Tether Gold
   XAUt = 'XAUt',
@@ -137,4 +141,5 @@ export enum CoinKey {
   WMON = 'WMON', // Monad Wrapped Native
   pBTC = 'pBTC', // Botanix Wrapped Native
   WTLOS = 'WTLOS', // Telos Wrapped Native
+  W0G = 'W0G', // 0G Wrapped Native
 }